South Database migrations for RockStor
Add south to settings.conf.in and setup.py so that South is installed
and added to INSTALLED_APPS.
Modify datastore and smartdb so that they contain the initial set of
tables for existing models (at the time of conversion), as well as the
south_migrationhistory table (when the app is converted to use south, it
fakes the initial migration, ie adds an entry to this table to show that
the app has been migrated to 0001).
Add migrations dir with initial migrations to storageadmin and
smart_manager.
Modify prep_db to check for the existence of data before creating it.
Modify buildout so that it
1. doesnt remove the databases.
2. runs migrations.
3. restarts servers.