Buildout modifications for database migrations
Add datastore.in and smartdb.in as templates to be copied if no database
exists.
Add stop-servers to shutdown servers.
Add clean-check section - if either datastore or smartdb don't exist,
assume its a fresh deploy, and copy the database files from the conf
dir.
Add db-migrate section to run migrations
In prep_db, check for existence of objects in a better way.
Move extra-stuff before collectstatic because collectstatic fails if the
log dir is not present