CPU widget resize handling
Add jquery.flot.resize plugin to handle resizing of graphs.
Remove row-fluid and span12 for the graph section of cpuusage template,
because the graph container is not able to figure out its height from
percentage values if they are present.
Set percentage heights and width on widget-content div, and the cpu
graph placeholder div, so that the resize plugin can automatically
resize acc to percentage value.