Fixes #80 Fixes #79
Enforce a minimum/maximum allowed size for a share during its creation in the share api
implementation. Make acceptable minimum and maximum values as config variables
in settings.py.

Ensure the above constraint is satisfied for share resize as well. In addition,
throw error when the new size given in the resize request is smaller than the
current usage of the share.