Making progress on detail views
After conferring with Suman, it was decided, after great debate that viewsets aren't the best choice at the moment. Ultimately,
every part of the URL's would be touched to place the appropriate viewset verb (list, create, retrieve, update). Instead, a new file
(by the name of detail_views) will hold the logic for any view that needs to retrieve a single object.