Separation of module and resource endpoints
In the new (>2.x) DRF, detail views and list views are strictly separated. This applies to POST
actions as well (in this instance, PUT and PATCH). For this, detail views were previously
constructed but it's now clear that they must contain much more code than originally thought (just
GETting).

This commit is the birth of that mindset. the rest_framework_custom/generic_view.py defaults to a
list view and a mixin will be used to apply the single object actions as needed (view this in
action in storageadmin/views/detail_views.py in DetailShareView).