Kernel version, uptime displayed
The event handler map is now used to invoke events as they come in from
the websocket connection. Very useful tool. Also, error handling for the
websockets was figured out. Essentially, to send useful error messages
to the client, you must use <code>self.emit()</code> but with a
different key. <code>self.error()</code> is useful for producing
messages on the server logs but not for translating information to the client.