Fixes to dashboard (null error), socket handler
The annoying error in the console from d3 is now fixed (something having
to do with transform=null). Also, the event handler is more robust. All
that's needed is RockStorSocket.addListener(callback, context,
'namespace'). The namespace is special because it expects a semicolon
separating the socket variable name as well as what the function
does (e.g. sysinfo:uptime).