Disable chronyd when ntpd is configured and turned on. Fixes #768
Chronyd is default in CentOS 7. It's a light weight alternative to ntpd
for non-server systems. It's on by default and serves the purpose well
for users that don't care about a robust time synchronization that ntpd
offers. However, if ntpd is configured, chronyd must be turned off to
keep ntpd on persistently across reboots. More info here:

https://access.redhat.com/documentation/en-US/Red_Hat_Enterprise_Linux/7/html/System_Administrators_Guide/ch-Configuring_NTP_Using_the_chrony_Suite.html