surround nut service toggle with try except & check for empty config #595
When the nut service is switched on without a prior config we have
a json type error in base_service.py. With this catch we present a
user friendly check config suggestion and do the same when other
config errors are encountered that stop the service from starting.
The underlying error is still logged.