move send_test_email to email_utils.
This is a temporary hack because email_client.py view file was
previously named email.py and because of lingering .pyc files, we run
into import problems. I was unable to cleanup .pyc files as part of
initrock or rpm spec file(oh, this only effects production style
deployment not development) so the final approach is to amend the
update_run to remove .pyc files(this will take effect in future) and
move any code that imports email to a directory outside of views
directory. with the update_run change in place, we can remove this
temporary fix in a future update.