retrieve ip of the current appliance dynamically. Fixes #917
Currently the ip field of an Appliance is static and is set when it's
created or a network interface is updated. But it's still same
information in two places in case of "current appliance". This change
makes it so that ip of the current appliance is dynamically retrieved to
be the management interface's ip or a random interface's ip in case no
interface is designated to be mgmt only.