remove quota rescan on share usage collection. Fixes #950
In 4.1 and IIRC 4.0 kernels, we had to issue quota rescans to refresh
the qgroup size information. This was resulting in high cpu usage at
regular intervals. In 4.2.1 kernel, the current one, it doesn't seem to
be necessary to rescan all the time. So the first step is to remove
these rescans.

Recent qgroup changes that went into 4.2 still need to be fully tested
in the context of Rockstor, but removing rescans is a good step in that
direction.