more logging to prove fix #1027
Fix is to return 1 as a general error condition as this is then
correctly interpreted by switch system as OFF and displayed
accordingly. -1 is interpreted incorrectly as ON which is
inaccurate and leads to inconsistent service status report.
N.B. normal stop status would return 3 so we still have
a distinction.