make missing / empty serial mechanism more robust #937
As we have recently enhanced the _update_disk_state mechanism with
the ability to accurately maintain removed drive info it is necessary
to enhance this flag mechanism that could previously under certain conditions
lead to repeat serial number entries in the db for removed devices. This in turn
could lead to a distortion of info on removed devices.
N.B. the previous UI flag mechanisms are maintained to allow for existing configs
that may have existing removed device info upon upgrade. All future flags will use
the new mechanism.