don't store removed fake-serial number drives #937
This doesn't make sense as they are not uniquely identified
and are give a fresh fake-serial number on each scan.
Without this a drive with no serial number will generate a
missing drive on each scan due to receiving a new fake serial
and hence causing the old fake serial to be registered as
now offline.