skip duplicate kernel dev names in scan_disks #1063
md* (multi device) block devices such as software raid
and some hardware raid devices are listed multiple times
in lsblk output, this causes them to be labeled as duplicate
devices due to shared serial number and any serial number
retrieved on the first occurrence is over-ridden on the second
occurrence by a fake to denote serial, and hence device,
duplication. This is incorrect in the case of these md devices.