generalize -u and env options to container. #858
Containers are run as root by default. Typically, the container profile
switches to a specific user and runs processes as that user. Some
containers take the uid of the container into account. Similarly,
ownership of mapped Shares can be taken into account. Make this
configurable in the Rock-on profile.

Handle env variables generically, similar to custom config, but at the
container level. To the end user it's completely transparent.