add handlebars comparison block helper for raw log length test #1084
An empty raw smart log (sata) is 4 lines long but the very
differently formatted scsi style non standard raw log is
more than 4 lines long. This block helper allows us to display
a log if it is longer than 4 lines which means it contains info
in both the sata and scsi cases.