in addition to chmod add chown on nut config files #1073
A background change in the OS / python has resulted in our
loosing the default nut.root ownership of nut's config files.
This breaks nut so as a matter of course we chown nut.root on
the files we edit. Previously we only needed the chmod as
nut group ownership was maintained. This is lost in some
installs, hence the requirement for chown now.