add role and smart_options fields to Disk model #1164
The role field allows us to tag drives when we currently only have
parted which is already overloaded. Used here to identify mdraid
members for pure and bios raid arrays. To cut down on model changes
during upgrades the smart_options was added at the same time; this
is to be used for custom smart parameters required for some drives:
mainly external usb caddy arrangements.