add specialized parser for in progress smart self test log entry #1207
Detect when we have failed to retrieve all 6 columns and if the cause
is down to in-progress test log entry then do additional parsing to
re-create our expected 6 columns. Cause is down to splitting on two
or more spaces and in this instance we have only one space between
columns. Special case so treat as such.