correctly retrieve previous spindown settings #885
As neither hdparm nor smart can tell us the current setting
of a drive as configured using -C (spindown time) we stash it
in a comment in the systemd file, it can then be used to pre-
select the dropdown for the same when re-configuring.
Thanks to @suman for assistance with this one.