tweak slider values as a temp work arount #885
This slider normally limits itself to the green and
yellow area, to work around this for now the values
have been tweaked to facilitate those we need.
Better later to code flag to not limit max to
(max - used) in d3.slider2.js. This works for now but
limits the visible red portion to a bare minimum.