add a 'remove config' option to the spin down config page #885
With this option we send a flag value of -1 (normally 1-255)
for the spindown time to indicate we want a current setting
(if any) removed rather than updated or added (if non exists).
On receiving this value we don't test the now nonsense hdparm
command, as it would make no sense, and go ahead and use the
existing mechanism to update our systemd file but this time by
removing an existing entry.