preserve existing drive role info on non mdraid devices #1214
As the role field was previously used only as an mdraid member
flag mechanism it would wipe all other roles, this was improved
in the case mdriad members having additional roles but all roles
would still be lost on non mdraid members. This is the initial
patch to address roles retention in non mdraid members.