if no by-id type disk name then fail over to caller name #1320
Disks with no serial number are not given a DEVLINKS entry by
udev ie virtio with no serial. This causes an issue as we need a
kernel name of sorts for every disk. Failing over to the calling
devices as found by scan_disks and flagging via an additional
return value (a boolean) associated via a tuple return type.
We can't use these no-serial devices anyway but do need to
have 'a' name in order to store and flag them.
Minor formatting updates and comment updates included.