disable SMART function for drives with transient names #1320
As we index the smart data to drive names and now these names
are less transient we have a chance to ensure that the SMART
data collected for a given drive will stick with that drive by using
only non transient names to index that data. So when we are
unable to retrieve a non transient name, ie when there is no serial
number and so no by-id name we disable the SMART data function.