minor rewording on disks_table to match detached device naming #1320
In the future the term offline in the context of btrfs may be more confusing
so for the time being at least move to attached / detached / reattached as
we then move closer to removable / backup / import type device roles.
Also detached device names are renamed to start with 'detached-' to
indicate their status in logs and current table views.