ensure the longest by-id name is used in get_byid_name_map() #1320
This removes the previous assumption that the longest by-id name for
a given device is always listed first as this was found to be incorrect.
Brings this function in line with the behaviour of get_dev_byid_name()
so as to maintain consistency. N.B. the context here is that some
devices have multiple by-id names so we take care to pick the longest
as it is more descriptive and gives repeatable naming.