add "or Notes" to errors column in balance table #1405
As this column is used for both errors and general messages
we should have the column title reflect this. An example of a
non error use is 'cancelled at x%' but we still want a status of
finished overall.
Also minor format changes (ie removal of tab char)

(cherry picked from commit eb8ec2d)