include 'cancelling', 'pausing' and 'paused' as blocker states #1022
Previously we only blocked raid changes on pools during balance
states of 'started' and 'running'. Due to recent changes we also need
to block on these additional balance states as seen in the db.
The 'paused' state may not strictly be a blocker but is probably
good practice to block in this state anyway.