Add pool usage widget to dashboard
https://forum.rockstor.com/t/usage-data-on-pools-and-shares-not-updated/423/24

A forum user wishes to have an accurate indication on the dashboard of
remaining storage space, to be able to tell when to add more. Currently
we only have a dashboard report of total storage capacity; this new
widget displays usage statistics per filesystem (pool). This is more
useful than per-share data given the current state of btrfs.