lead SMART attributes table with Name column #1001
Moved the Attributes Id column from the beginning to
the end of the table. As this is a human view it makes
sense to lead with the now alphabetically sorted name
column rather than the id column, however reference
via the id column is still afforded by the edge table position.