Make top level mount directories read-only when unmounted. Fixes #1414
Shares and Pools are mounted as /mnt2/<name>. Sometimes they may fail to
mount but applications write to these directories assuming the writes
are going to a Share. But rootfs fills up instead and can cause major
problems later on. To fix this problem, set the read-only
attribute(chattr +i) on Pool, Share and export directories when they are
created. With this change, these directores stay read-only unless the
corresponding Share, Pool or nfs export is mounted.

Sort constant declarations.

Remove Shares properly when force deleting the parent Pool