add front end logic to use disk.role to identify bcache devices #1494
We differentiate between backing devices and caching devices as their
purpose is very different ie hdd's for backing devices and much faster
devices such as ssd for caching devices. The distinction helps to confirm
correct configuration. A double link icon is used to identify cache devices
and a single link for backing devices.