add initial validation to disk role selection UI #1494
We need to disallow redirection away from a btrfs
partition as we favour these and it will break pools if
this is done. So disable redirects if there is an existing
btrfs partition unless it is to select this partition. And
once a btrfs part is selected for a redirect role refuse
to remove it whilst the btrfs fs exists.