use 'tags' icon in all device tables to indicate a redirect role #1494
It is useful to know if a pool member is partitioned (using a redirect
role). If so label them in a manner consistent with the main disks page
with context relevant tooltips and links to review / edit the redirect
role. Note: as the 'root' pool partition member is not managed by
a redirect role, exclude it from this iconic flagging.