back propagate crypto_LUKS partition info to the base device #550
In order to appropriately identify a disk as hosting a LUKS container
within one of it's partitions we label the base device with this canonical
indicator. This is the same treatment used to identify mdraid disk
members when any one of their partitions is an mdraid member.
Predominantly used to accommodate default anaconda system disk
encrypted (LUKS) installs (ie swap and btrfs system pool partitions
as LUKS containers).