refactor and enhance role flags UI code on Disks page #550
Reduces the number of redundant disk status checks re role
and state of each disk and improve readability and logic
of the associated checks. The intention is to fashion a user
'funnel' to accommodate early LUKS / bcache device role
assignment while attempting to reduce the number of
redundant or unwise configuration options offered. I.e. on
freshly wiped devices a LUKS / bcache tooltip suggestion
introduces the UI entry point to apply such whole disk
configurations.  Note also that a new flag is introduced, that
of a drive mapping (essentially if the device is found to be part
of a pool) this is intended to aid in identifying fresh or already
pool committed drives from the flags alone and works in
concert with LUKS / bcache flags, ie one will replace another.