initial commit of LUKS UI page #550
We already have a LUKS format extension on the roles
page but that is akin to the existing wipe / delete
function found there already. This page is intended
as the home for existing LUKS containers and open
LUKS volumes where we might put such utilities as
luksHeaderBackup, passphrase facilities etc. Also
the ability to open / close a given container /
volume respectively.