remove LUKS format option on LUKS containers #550
When we view LUKS containers on the roles / wipe page
it can be misleading to offer to LUKS format them, it
may lead to a misunderstanding about if the last format
'took'. I.e don't LUKS format when already LUKS
formatted. Also we open the way to use the existing wipe
mechanism to remove existing LUKS containers which helps
to keep the wipe mechanism, and it associated validation,
centralized.