add 'setup' delay for bootstrap APIWrapper #1422
On 2 sets of roughly 10 consecutive reboots, one set
pre and one set post patch, the prior arrangement (no
delay) took 2 'tries' to complete successfully on 100%
of the boots; ie always failed on the first attempt.
After the additional delay was introduced 100% of the
boots did not produce an exception: completing the
'try' block on the first attempt.