extend volume_usage() unit test for pqgroup=-1/-1 #1769
In some instances the share.pqgroup db field can contain
the nonsensical '-1/-1'. This unit test extension adds a
call to volume_usage() with this value. This is to prove
a pending fix to volume_usage() and help guard against
future regressions of the same type: volume_usage()
failing to return the expected number of list elements
when passed a non existent qgroupip as 3rd parameter.