pool name in tasks overview table as link #1243
As we already indicate the pool (by name) that a
scheduled scrub task relate to it is convenient to have
this link to that pools details page. N.B. todo added
for the same to be done re share names and snapshots:
pending share API change / fix in tasks. Plus minor tab
char removal.