return realtime info in task defition via properties. Fixes #1857
information like share_name and pool_name can be supplied when the model is
serialized during a GET, by constructing it in real time. This pattern is
already there in a few other models and works nicely for any information that
shouldn't be stored in db or at least not duplicated. The other way of solving
this problem is to have UI/client make explicit calls to appropriate end points
to get share_name etc.. which is more RESTful, but introduces unnecessary
complication in our case.