version and date incorrectly reported re update info #1870
A recent move to yum from rpm inadvertently caused a miss
reporting of installed version for that of available version.
This was due to the differing outputs and the inclusion of
both installed and available in the newer yum format.
The command switch set was revised to output only the installed
version and to include the previously missing and required build
date/time info. Consequent parsing adjustments were made to
accommodate for the now changed date format.
Time info was excluded as it is not used.