improve quotas not enabled behaviour #1869
Uses the existing -1/-1 default for pqgoups to represent
an unset state for the share.pqgroup and adds an active
return to this value whenever quotas are deemed to be
disabled. Pqgroup setup is moved from the bootstrap
process into the import_shares / shares refresh section.
This allows for live setting / resetting of pqgroups via
the regular share re-fresh process. Pool.quotas_enabled
and share.pqgroup_exist properties are added and all used
low level quota actions are adjusted to catch, log, and
ignore a quota disabled state.
Additionally active pqgroup assignment is added to share
resize and share refresh procedures which aid in returning
to existence the expected native 2015/n pqgroups and their
relationship to the auto generated 0/n qgroups as whenever
quotas are disabled all this info (var the 0/n) is lost.
UI elements are added in pool and share focused pages to
indicate a live status for the associated pool quotas
enabled status.