Fix replication regression re share api change #1853
Prior function was restore by updating the relevant api urls in the
replication system. But in one, non critical check this was not managed.
As a result the non critical but desirable internal sanity check was,
or the time being, remarked out and TODOs added to signify required
future attention. Also includes a number of UI fix ups including table
formatting, sort capability, and consistency with the rest of the UI.

A core mechanic of the replication system was also abstracted such that
it could benefit from the existing share / snapshot management system.
This alteration made more explicit, in code, a quirk involving the
initial subvol transferred and how it requires a special treatment
given it's unique nature in the context of the existing share /
snapshot / clone / import structures.

- Update internal replication api urls to fit new share by id scheme.
- Remove problematic non critical internal check (TODO added for later).
- Create new repclone command to improve replication share / snap
handling, ie more transactional by using existing structures via api.
- Minor improvements were added in the new repclone system, ie sys
refresh qgroup transfer, auto mounting etc.
- Ensure we import initial btrfs receive quirk subvol with replica flag
and add logging to indicate it's quirk nature.
- Improve replication tables formatting, text, sort ability / order.
- Reduce and mirror (between send and receive) rep-snap-count.
- Fix bugs in receive status update, includes utilising the pending
state.
- Improve user communication re data transferred / rate table cell text.
- Remove broken pool link in replication overview table.
- Improve debug logging re share / snap import.
- Minor steps / TODOs towards future share name/subvol_name seperation.
- TODOs towards centralising mount path creation.
- Various TODO added for future consideration.