fix clone auto mount on creation #1836
Previously when creating a new share via the clone
mechanism from either an existing share or a writable
snapshot the resulting share would not be auto mounted.
This is not consistent will all other 'new' share scenarios.
Also includes a qgroup / pqgroup fix where the resulting
share was not assigned a rockstor native 2015/x pqgroup
and update_quota was incorrectly called on the auto created
0/x qgroup rather than the pqgroup.