remove immutable flag prior to share delete #1882
As an immutable flag can break share delete and given
we already provide a UI warning, this preemption of
and existing immutable flag is deemed appropriate.
Especially in the context of potentially breaking
replication. It is however advised to re-address this
approach once the sporadic nature of immutable flags
appearing without manual intervention is resolved.