allow import of ro,degraded cli mounted pool #1890
Facilitates the import of a prior cli mounted pool, this enables
the ability to use initial mount options that are currently
not available via the Web-UI prior to an import. The focus here
is to enable an import of a pool that requires, for example,
ro,degraded mount options.

1. Fix import failure (prior TODO:) via Read-only quota enable attempt
on import.
2. Fix failure on qgroup assign on ro fs and avoid these and related
calls from higher level functions via (4) below.
3. Improves code consistency re PQGROUP_DEFAULT use.
4. Make additional use of the existing quota disabled mechanism to
improve ro behaviour during import and subsequent share/snap refresh.
5. Avoid assigning pqgroups within db that failed creation due to ro fs.
6. Improve quota disabled robustness re redundant e.rc checks from btrfs
command execution as per a previous pr code review.
7. Improve / update code comments.
8. Improve pqgroup/quota management.