surface pool missing disk info in UI #1897
By adding a lighweight 'has_missing_dev' property to the Pool object we
can conveniently surface pool degraded status within the UI. The pools
overview and detail pages can then be informed by this property and
consequently display red "Some Missing" indicators within the pool's
respective disks section. In addition another mechanism linked to a new
gevent is used to inform system wide pool degraded status to the Web-UI
header, so that no page refresh is required. This second mechanism can
also report on the number of non managed / pre-imported pools. The
extraction of pool labels for unmanaged pools is purposefully avoided
in order to keep the initial detection algorithm as light as possible.
On systems with no degraded pools the detection is down to a single
system call. Pool Object creation for the header gevent notifications
is avoided until at least one degraded pool is detected.

Summary:

- Add serialized 'has_missing_dev' boolean property to Pool Object.
- Add lightweight pool degraded detection to inform above.
- Add red "SOME MISSING" UI element to Disks column in Pool overview.
- Add red "Some Missing" UI element to Disks title in subsection of Pool
details view.
- Add system wide pool degraded detection to inform Web-UI header, via
a gevent, so that even on a freshly installed system, unimported pools
can be indicated. This system does not require a Web-UI refresh and is
modeled after the existing 'shutdown-status' indicator.
- Add a unit test to prove the function of the above system wide pool
degraded sensor.
- Extend the existing back-end pool-info system to encompass degraded
state along with expected total disk count and missing disk count. These
additional facilities are not employed in this pr but will serve as a
double check on Pool object info to inform future mechanisms before
enacting on such operations as delete missing.
- Add TODO on informing system pool's status.