unmounted pool with prior balance record has broken details page #1837
When visiting an unmounted and un-mountable pool's details page and that
pool has an existing balance record, there is an attempt to update that
record with a current balance status assessment. This in turn initiates
a pool mount attempt; and when this mount attempt fails the exception is
not caught and is surfaced: breaking that Pool's details page.

Summary:

- Catch and log pool mount exception in balance_status(): returning
'unknown' as without a mounted fs we can't know the balance state.
- Add unit test for new mount exception behaviour in above.
- Add unit test for non exception 'unknown' balance_status().
- Address prior TODO: on missing balance_status() unit test: 'finished'
status.
- Cosmetic / consistency enhancements to existing balance_status() unit
tests.