add root subvol exclusion mechanism #1931
Our current share/clone subvol surfacing mechanism assumes a top level
dir exists for a given pool (vol). This assumption is most notably
flawed in the case of a system root ('/') as by definition it is the
top level and has no '/' + 'dir-name' counterpart. However our current
default system subol arrangement has '/' in a subvol named 'root' which
coincidentally has the FHS counterpart subdir of '/root' and current
btrfs commands return similarly for a subdirectory if that subdirectory
is not also a subvol in it’s own right, which '/root' in our case is
not.  The mechanism chosen to initially address this issue is to
exclude the Rockstor native re-mounting (in /mnt2) of known root
subvolumes: ie: 'root', '@' and or subvols configured to be the default
pool (vol) mount point (btrfs subvol get-default /) ie.
'/@/.snapshots/1/snapshot' (ie a default snapper root config).

The subvol exclusion mechanism, as instantiated in this commit/pr, also
serves to surface subvols of excluded subvols; ie '@/home' where @ is
itself a subvol: a more common arrangement in modern root on btrfs
configurations. This is due to our current share/clone single depth
subvol consideration.

A wrinkle in subvol path exclusion practice is that when the '/'
subvolume does not share a direct common subvol as parent, such as in a
default snapper root config boot to snapshot arrangement, the path
expressed for '/home' is then eg '@/home' as opposed to our previously
expected and relative to shared direct parent (@) 'home'. This is
accounted for, in this commit's / pr's first pass treatment, by simply
doubling up relevant exclusions and hard wiring a blind stripping of
leading '@/' chars: a sub optimal but proven functional (by included
unit tests) fix. It is suggested that this approach be iteratively
improvement upon.

Summary:
1. Establish a named root subvol exclusion list.
2. Add a function to identify the default subvolid for the pool hosting
‘/’.
3. Reference the above 2 mechanisms to 'skip' consideration of
associated subvols, by path and or by id respectively by way of (1.)
and (2.) above (debug logged).
4. Refactor shares_info() (creating get_property() and snapshot_idmap())
to help remove code duplication and improve code clarity clone
identification logic, and testability.
5. Add a redirect to '/' in cases where we are examining an excluded
subvol which has no /mnt2 counterpart.
6. Brute force conversion to currently recognized relative paths, ie
strip @/ chars.
7. Normalise use of toggle_path_rw() for immutable flag manipulation.
8. Remove redundant/unused function, See Note * for details.
9. Removed an outer scope variable name collision.
10. Account for upstream parent-child column inversion -
group_is_assigned().

Note that 10 in above is an unrelated change to issue #1931 but is
trivial in nature: upstream issue reference included in code comments.

Note *
Removed subvol_list_helper() and sole remaining user. Introduced 4 years
8 months ago in #105 to mitigate an rc=19 subvol list issue. This
helper was not universally employed and we have had no reports of it's
recent occurrence; hence removal. Comment left by prior sole use
location, and related remaining rc = 19 exception clause.