Update javascript templates for new Device choice form
Create new templates for Device choice modeled after the "Custom_choice" ones, with the deletion of the "required" CSS class and asterisk depiction on the form fields.
Also attempt to update the Summary and Install Summary tables to show the newly devices as well. Unfortunately, as it is the case for environment variables, the device choices do not seem to show on any of those tables, despite being passed successfully to docker run command.