regression in dashboard disk activity widget #1978
Recent improvements / bug fixes in the non dashboard by-id name
selection has lead to a deviation in algorithms for by-id name
resolution leading to a failure in the disk activity widget due to
unknown by-id name selection within that widget. This only happens if
all by-id names for a given device are of the same length.

Summary:

- Move dashboard 'temp name -> by-id' map generation to reverse
lexicographical to normalise on by-id name selection across the system.
- Add TODO on potential dashboard code simplification in this area: has
performance test pre-requisite.
- Add unit tests for both prior and current behaviour of
get_byid_name_map() to ensure nature of change is understood.