Remove ancient 64-btrfs.rules so systemd provided rule is used
This solves the failure to boot when the root fs is on a btrfs handled
raid set-up. The reason it fails is typically because one or both disks
failure are not initialised before the kernel wants to mount root.

This generally affects only disks that are mounted during boot.

Solves issues #1329 and #1980