pin python-engineio to 2.3.2 as recent 3.0.0 update breaks gevent #1995
As of 25th November 2018 source builds failed to display live Web-UI
header info such as time, Rockstor and kernel version etc. This was
tracked down to an upstream python-engineio updated to 3.0.0. Although
this update may not be the direct cause, pinning it's version to the
latest prior released resolved the observed regressions.