change default return value of get_global_config for samba
When working with Samba, if you navigate to the dedicated page for it, we internally save the result of get_global_config in the model.
This then causes subsequent checks to open the configuration modal to fail, since it's expecting null rather than an empty dictionary.
This change simply uses the default of None, and then swaps to a dictionary as needed.