remark out all currently failing unit tests #2137
All currently failing unit tests on our to-be-legacy but still reference
CentOS source build are remarked out in order that we can more readily
use the existing tests as indicators of disparity between our ditro
bases. In almost all cases the fail/error message of the test is noted
in TODO's. The vast majority failures are suspected as being caused by
a lack of maintenance within the tests and several major changes such as
disk/share/pool/network API changes

Includes:
- Remark out 2 sftp sub tests, broken/out of sync with core code.
- Remark out test_btrfs_disk_import: 500 not 200.
- Fix broken test_commands.py CommandTests - core refactor bit rot. see:
"[openSUSE] use zypper for package installs. Fixes #2071" (pr #2081)
in 3.9.2-51. The unrelated get_pool_info() breakage was not tracked
down.
- Update patch targets due to core refactoring. In pr: "Add unit testing
for core network functions" #2045 at least a couple of functions were
renamed. Updated these in our network API test setup.
- Remark out all network API tests as this API changed long ago.
- Remark out many whole and sub-parts of the nfs tests. Many of the
failures here are likely down to disk / share API changes after these
tests were written. Specifically Shares API in this case as there are
many (noted in TODO: comments): "share instance ... does not exist."
which suggests a failure of the mocking arrangement.
- Remark out test_share_acl as currently non working and need work.
- An attempt was made to fix the mocking required to have this test work
but it was not a success. Minor minification were left in place
never-the-less.
- Remark out test within test_oauth_apply - Needs updating.
- Update to account for refactoring in tested & add new needed mocks.
- Since this test was update there have been a few pool model changes.
The added mocks mostly move this test back to where it was prior to the
pool management code changes.
- Remove overly ambitious/broad test_raid_migration - failing.
- Fix missing space in expected error message test_raid6_crud.
- Remark out currently unworkable sub tests within test_raid6_crud.
- Our pool / disk member mocking is currently insufficient and so a
range of tests now fail. Remove for to use what we can.
- Remark out currently unworkable sub tests within test_raid1_crud.
- Our pool / disk member mocking is currently insufficient and so a
range of tests now fail. Remove for to use what we can.
- Remark out several sub tests within test_samba.py as need fixing.
- Remark out a sub test within test_group as need fixing or OOD.
- Remark out a sub test within test_login as non representative.
- Remark out a sub test within test_shares as incorrectly triggered.
- Remark out all but on test in test_oauth as test seem broken.
- Remark out one more test in test_snapshots as known to be duff.
- Remark out several more test in test_user as known to be duff.