BTRFS SCRUB status enhancements for rockstor
Updated the scrub status code to check the version of btrfs-progs that is installed. Starting with 5.1.2, the scrub status output changed. This code will handle both cases, either usin ghte newer output format or the older.

Additionaly, if running with the newer btrfs progs, modified the Scrub Status Summary and Detail pages to show the ETA as well as the current scrub speed.