add observability clause for subvol of snapshot for boot to snap #2156
By refactoring our existing call to default_subvolid() and introducing
an additional clause for subvols that we do surface to the Web-UI we fix
the 'disappearing' system pool user share on page refresh and reduce the
total number of calls to default_subvolid(). An additional btrfs unit
test was added for this additional clause based on real boot to snapshot
data.