add mount point instance variable/property to Pool & Share models #2156
Instantiating the mount point for Pool related operations at the
Pool/Share model level allows for contextual treatment of the system
share which in turn facilitates the removal of the prior double mount of
'/': once as '/' via fstab and again as /mnt2/ROOT. And given the latter
was our prior mount point of pool related operations, where some btrfs
commands fail on boot-to-snapshot configurations, we return full
function, as per non boot-to-snapshot configs, by addressing all pool
related btrfs operations (ie some subvol and quota related calls) to the
otherwise functional primary mount point (fstab managed) of '/'.
Effectively fixing a host of boot-to-snapshot issues that arose re a
subset of btrfs calls and removing the redundant, in the case of the
system pool, /mnt2/ROOT mount point.